requirements.txt
bcrypt==4.0.1
fastapi[standard]
pyjwt
passlib
pydantic
mysql-connector-python